fixed issues with saving on service start/stop

+++ temperatureMonitor.py
@@ -6,8 +6,12 @@ from si7020_I2C import si7020_I2C
 import numpy as np
 import matplotlib.pyplot as plt
 import os
+import signal
 from datetime import datetime, timedelta
 import atexit
+
+# temperature and Humidity
+TandH_data = []
 
 def sleep_until_next_minute():
     current_time = datetime.now()
@@ -20,7 +24,7 @@ def sleep_until_next_minute():
     time.sleep(sleep_seconds)
 
 
-def plot_and_save_data(data):
+def plot_and_save_image(data):
     timestamps, humidities, temperatures = np.transpose(data)
 
     # Convert temperatures to numerical values
@@ -71,8 +75,8 @@ def plot_and_save_data(data):
 def check_time(save_interval):
     timestamp = datetime.now()
     # Check if it's within the time window to save data
-    target_time_start = datetime.strptime('02:00:00', '%H:%M:%S').time()
-    target_time_end = (datetime.strptime('02:00:00', '%H:%M:%S') + timedelta(seconds=save_interval)).time()
+    target_time_start = datetime.strptime('11:40:00', '%H:%M:%S').time()
+    target_time_end = (datetime.strptime('11:40:00', '%H:%M:%S') + timedelta(seconds=save_interval)).time()
     current_time = timestamp.time()
     if target_time_start <= current_time <= target_time_end:
         return True
@@ -99,17 +103,35 @@ def create_time_encoded_file_name(name):
 
 def save_data_to_file(data, filename):
     """ same temperature and humidty as a text file for future"""
-    with open(filename, 'w') as file:
-        # Write header
-        file.write('Timestamp,Humidity (%),Temperature (F)\n')
+    if os.path.exists(filename):
+        # File exists, load data into list 'data'
+        with open(filename, 'w') as file:
+            # Write header
+            file.write('Timestamp,Humidity (%),Temperature (F)\n')
 
-        # Write data
-        for data_point in data:
-            formatted_data_point = [f'{val:.3f}' if isinstance(val, (float, float)) else str(val) for val in data_point]
-            file.write(','.join(formatted_data_point) + '\n')
+            # Write data
+            for data_point in data:
+                formatted_data_point = [f'{val:.3f}' if isinstance(val, (float, float)) else str(val) for val in data_point]
+                file.write(','.join(formatted_data_point) + '\n')
+    else:
+        file_not_found_file_name = f"temperatureMonitor_data_{datetime.now().strftime('%Y-%m-%d')}.txt"
+        with open(file_not_found_file_name, 'w') as file:
+            # Write header
+            file.write('Timestamp,Humidity (%),Temperature (F)\n')
+            # Write data
+            for data_point in data:
+                formatted_data_point = [f'{val:.3f}' if isinstance(val, (float, float)) else str(val) for val in data_point]
+                file.write(','.join(formatted_data_point) + '\n')
 
 
 def main():
+    global TandH_data
+    #syslog every 15 minutes
+    report_time = (datetime.now() + timedelta(minutes=15)).replace(second=0, microsecond=0)
+    # save at mid nite unitl I run this as a script from allsky
+    save_time = datetime.now().replace(hour=0, second=0, minute=0, microsecond=0) + timedelta(days=1)
+
+    # gpio controls
     pi = pigpio.pi()
     th_sensor = si7020_I2C()
     if not pi.connected:
@@ -124,24 +146,24 @@ def main():
     while True:
         # Call temperature and humidity functions
         humidity, temperature = th_sensor.get_humidity_and_temperatureF()
-        timestamp = datetime.now()
-        print(" time is ", timestamp.strftime('%Y-%m-%d %H:%M:%S'))
         # Round values to one significant digit
         temperature_rounded = round(temperature, 1)
         humidity_rounded = round(humidity, 1)
 
         # temp humidity and timestampe as its own data file
-        data_point = [timestamp.strftime('%Y-%m-%d %H:%M:%S'), humidity, temperature]
-        data.append(data_point)
+        time_stamp = datetime.now().strftime('%Y-%m-%d %H:%M:%S')
+        data_point = [time_stamp, humidity, temperature]
+        TandH_data.append(data_point)
+
         # at midnight we start a new file
-        if check_time(save_interval):
-            save_data_to_file(data, todays_name)
-            # Clear data for the new day
-            data = []
+        if save_time < datetime.now():
+            print(f"hello steve, save time:{save_time} time now:{datetime.now()}")
+            save_data_to_file(TandH_data, todays_name)
+            TandH_data = []
+            save_time = datetime.now().replace(hour=0, second=0, minute=0, microsecond=0) + timedelta(days=1)
         else:
             # Create and save plot
-            print("saving data")
-            plot_and_save_data(data)
+            plot_and_save_image(TandH_data)
             todays_name = create_time_encoded_file_name("_temperature_humidity_data")
 
         # now check fan and heater status
@@ -161,24 +183,28 @@ def main():
 
         # now put in json format and save for overlay to use
         sensor_data = SensorData(temperature_rounded, humidity_rounded, heater, fan)
-        sensor_data.data_dict["HS_TIME_STAMP"] = timestamp
+        sensor_data.data_dict["HS_TIME_STAMP"] = time_stamp
         # Update data dictionary
         sensor_data.save_to_json(json_file_path)
 
-        syslog.syslog(syslog.LOG_INFO, f"Allsky Temperature: {temperature_rounded} F")
+        if report_time < datetime.now():
+            report_time = (datetime.now() + timedelta(minutes=15)).replace(second=0, microsecond=0)
+            syslog.syslog(syslog.LOG_INFO, f"Allsky temperature: {temperature_rounded} F")
 
         # Sleep for a minute
         sleep_until_next_minute()
 
-def shutdown_save():
+def shutdown_save( signal, stack_frame):
     todays_name = create_time_encoded_file_name("_temperature_humidity_data")
-    save_data_to_file(data, todays_name)
+    save_data_to_file(TandH_data, todays_name)
     syslog.syslog(syslog.LOG_INFO, f"temperature monitor shutting down saving data to: {todays_name}")
+    exit(0)
 
 
 atexit.register(shutdown_save)
+signal.signal(signal.SIGTERM, shutdown_save)  # SIGTERM is the default signal sent by systemctl stop
+signal.signal(signal.SIGINT, shutdown_save)   # SIGINT is sent when you press Ctrl+C
 
-data = []
 
 if __name__ == "__main__":
     main()
